% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{age_predict}
\alias{age_predict}
\title{Прогнозирование возраста указанного пользователя}
\usage{
age_predict(user_id = "")
}
\arguments{
\item{user_id}{Идентификатор пользователя, для которого необходимо определить возраст}
}
\description{
Прогнозирование возраста указанного пользователя
}

