% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{areFriends}
\alias{areFriends}
\title{Checks the friendship status between two users}
\usage{
areFriends(source_id, target_id)
}
\arguments{
\item{source_id}{Source user ID}

\item{target_id}{Target user ID}
}
\description{
Checks the friendship status between two users
}
\examples{
\dontrun{
areFriends(me(), 123456)
}
}

