% Generated by roxygen2 (4.0.2): do not edit by hand
\name{authorize}
\alias{authorize}
\title{Авторизация и получение ключа доступа}
\usage{
authorize(client_id = "", scope = "", email = "", password = "")
}
\arguments{
\item{client_id}{Идентификатор приложения}

\item{scope}{Битовая маска настроек доступа приложения, которые необходимо проверить при авторизации пользователя и запросить, в случае отсутствия необходимых.}

\item{email}{Номер телефона или почтовый адрес}

\item{password}{Пароль}
}
\description{
Авторизация и получение ключа доступа
}
\details{
Ниже перечислен список некоторых из возможных прав доступа, полный список может быть найден на
странице документации \href{https://vk.com/pages?oid=-1&p=Права_доступа_приложений}{Права доступа приложений}:
\itemize{
  \item \strong{friends} Доступ к друзьям.
  \item \strong{photos} Доступ к фотографиям.
  \item \strong{audio} Доступ к аудиозаписям.
  \item \strong{video} Доступ к видеозаписям.
  \item \strong{docs} Доступ к документам.
  \item \strong{notes} Доступ заметкам пользователя.
  \item \strong{pages} Доступ к wiki-страницам.
  \item \strong{status} Доступ к статусу пользователя.
  \item \strong{wall} Доступ к обычным и расширенным методам работы со стеной.
  \item \strong{groups} Доступ к группам пользователя.
  \item \strong{messages} Доступ к расширенным методам работы с сообщениями.
  \item \strong{notifications} Доступ к оповещениям об ответах пользователю.
}
}

