% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_count.R
\name{batch_count}
\alias{batch_count}
\title{Print summary statistics for count models}
\usage{
batch_count(
  df,
  outcome,
  predictor,
  covariates = NULL,
  numdec = 2,
  verbose = FALSE
)
}
\arguments{
\item{df}{Input data frame}

\item{outcome}{Outcome measure}

\item{predictor}{Predictor}

\item{covariates}{Single covariate or array of covariates}

\item{numdec}{Number of decimals for the beta and std statistics}

\item{verbose}{Set to TRUE if you want to see the full lm output}
}
\description{
This function prints summary statistics for a count models such as the
estimate, std.err, p-value, and effect size for a predictor. It will
automatically detect if the predictor variable is numeric or a factor
variable. If it is a factor variable, then estimates etc. will be provided
for all levels of the facor variable (except the first level, that will be
the reference category).
}
\details{
It will also detect if a Poisson model or negative binomial regression model
is a better fit fo the data, and will report the information for the most
appropriate model. Model selection is done by running both models and then
using a likelihood ratio test to evaluate if either Poisson or negative
binomial is the better fit.

References:
\enumerate{
\item https://stats.oarc.ucla.edu/r/dae/negative-binomial-regression/
}
}
