% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_lm_fdr.R
\name{batch_lm_fdr}
\alias{batch_lm_fdr}
\title{Apply FDR correction to batch_lm output}
\usage{
batch_lm_fdr(idf)
}
\arguments{
\item{idf}{Input data frame}
}
\description{
This function takes output from batch_lm and grabs all the values in the
columns ending in _fdr. batch_lm stores full length (not truncated) raw
p-values there. It creates one array of these p-values and then applies FDR
correction to it. After that, it stores the corrected p-values back into
these columns.
}
