% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_lmer_lmer.R
\name{batch_lmer}
\alias{batch_lmer}
\title{Print summary statistics for a linear mixed-effects model}
\usage{
batch_lmer(df, outcome, predictor, reffect, covariates = NULL, verbose = FALSE)
}
\arguments{
\item{df}{Input data frame}

\item{outcome}{Outcome measure}

\item{predictor}{Predictor}

\item{reffect}{Random Effect variable}

\item{covariates}{Single covariate or array of covariates}

\item{verbose}{Set to TRUE if you want to see the full lm output}
}
\description{
This function prints summary statistics for a linear model such as the
estimate, std.err, p-value, and effect size for a predictor. It will
automatically detect if the predictor variable is numeric or a factor
variable. If it is a factor variable, then estimates etc. will be provided
for all levels of the facor variable (except the first level, that will be
the reference category)
}
\details{
References:
\enumerate{
\item https://github.com/bcjaeger/r2glmm
\item https://stats.stackexchange.com/questions/358927/compute-partial-eta2-for-all-fixed-effects-anovas-from-a-lme4-model
}
}
