% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_lmer_wrapper.R
\name{batch_lmer_wrapper}
\alias{batch_lmer_wrapper}
\title{Run batch_lmer and subsquently batch_lm_fdr on a number of predictor and
outcome variables.}
\usage{
batch_lmer_wrapper(
  df,
  outcomes,
  predictors,
  reffect,
  covariates = NULL,
  numdec = 2,
  glmmnb = FALSE,
  fdr = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{Input data frame}

\item{outcomes}{Outcome measure(s) to loop over}

\item{predictors}{Predictor(s) to loop over}

\item{reffect}{Random Effect variable}

\item{covariates}{Single covariate or array of covariates}

\item{numdec}{Number of decimals for the beta and std statistics}

\item{glmmnb}{Set to TRUE if you want to use a Negative Binomial GLMM}

\item{fdr}{Apply FDR correction (default = TRUE). Turn it off if you want to merge multiple tables later and apply FDR correction on the combined table.}

\item{verbose}{Set to TRUE if you want to see the full lm output}
}
\description{
This function is a wrapper script for batch_lm to loop over a number of
predcitor and outcome variables, and to subsequently apply FDR correction on
the output using batch_lm_fdr.
}
