% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{boardGetComments}
\alias{boardGetComments}
\title{Returns a list of comments on a topic on a community's discussion board}
\usage{
boardGetComments(group_id = "", topic_id = "", need_likes = 0,
  start_comment_id = "", offset = 0, count = 20, sort = "",
  extended = 0, v = getAPIVersion())
}
\arguments{
\item{group_id}{ID of the community that owns the discussion board.}

\item{topic_id}{Topic ID.}

\item{need_likes}{1 - to return the likes field, 0 - not to return the likes field (default).}

\item{start_comment_id}{Positive number.}

\item{offset}{Offset needed to return a specific subset of comments.}

\item{count}{Number of comments to return (default 20, maximum 100).}

\item{sort}{Sort order: asc - chronological, desc - reverse chronological.}

\item{extended}{1 — to return information about users who posted comments; 0 — to return no additional fields (default).}

\item{v}{Version of API}
}
\description{
Returns a list of comments on a topic on a community's discussion board
}

