% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{boardGetCommentsExecute}
\alias{boardGetCommentsExecute}
\title{Returns a list of comments on a topic on a community's discussion board}
\usage{
boardGetCommentsExecute(group_id = "", topic_id = "", need_likes = 0,
  start_comment_id = "", offset = 0, count = 20, sort = "",
  progress_bar = FALSE, v = getAPIVersion())
}
\arguments{
\item{group_id}{ID of the community that owns the discussion board.}

\item{topic_id}{Topic ID.}

\item{need_likes}{1 - to return the likes field, 0 - not to return the likes field (default).}

\item{start_comment_id}{Positive number.}

\item{offset}{Offset needed to return a specific subset of comments.}

\item{count}{Number of comments to return (default 20, 0 - for all comments).}

\item{sort}{Sort order: asc - chronological, desc - reverse chronological.}

\item{progress_bar}{Display progress bar.}

\item{v}{Version of API}
}
\description{
Returns a list of comments on a topic on a community's discussion board
}
