% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{boardGetCommentsList}
\alias{boardGetCommentsList}
\title{Returns a list of comments on a community's discussion board}
\usage{
boardGetCommentsList(topics, progress_bar = FALSE, v = getAPIVersion())
}
\arguments{
\item{topics}{A list of topics (from getTopicsExecute())}

\item{progress_bar}{Display progress bar}

\item{v}{Version of API}
}
\description{
Returns a list of comments on a community's discussion board
}

