% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clear all variables in the workspace}
\usage{
clear(exceptions)
}
\arguments{
\item{exceptions}{A variable or list of variables that you want to keep}
}
\description{
This function allows you to clear all variables
in your works space and to select specific variables
that you do not want to be removed.
}
