% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_text}
\alias{clear_text}
\title{Clear text}
\usage{
clear_text(lines, patterns = list())
}
\arguments{
\item{lines}{List of lines}

\item{patterns}{List of user defined patterns}
}
\description{
Clear text
}

