% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{collection_exists}
\alias{collection_exists}
\title{Check if collection exists}
\usage{
collection_exists(collection_name, collection_suffix = "",
  db_name = db_getActive())
}
\arguments{
\item{collection_name}{Collection name}

\item{collection_suffix}{Collection suffix}

\item{db_name}{Database name}
}
\description{
Check if collection exists
}
