% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{create_empty_collection}
\alias{create_empty_collection}
\title{Create empty collection}
\usage{
create_empty_collection(collection, suffix, db_name = db_getActive())
}
\arguments{
\item{collection}{Collection name}

\item{suffix}{Collection suffix}

\item{db_name}{Database name}
}
\description{
Create empty collection
}

