% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropplot.R
\name{cropplot}
\alias{cropplot}
\title{Crop a pdf plot using inkscape}
\usage{
cropplot(figfile)
}
\arguments{
\item{figgile}{The pdf file you want to crop.}
}
\description{
This function removes all white space around a pdf figure. The input file
will be overwritten.
}
