% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetChairs}
\alias{databaseGetChairs}
\title{Returns list of chairs on a specified faculty}
\usage{
databaseGetChairs(faculty_id = "", offset = "", count = "100")
}
\arguments{
\item{faculty_id}{ID of the faculty to get chairs from.}

\item{offset}{Offset required to get a certain subset of chairs.}

\item{count}{Number of chairs to return.}
}
\value{
Returns a list of chairs objects with id and title fields.
}
\description{
Returns list of chairs on a specified faculty
}
\examples{
databaseGetChairs(206)
}

