% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetCities}
\alias{databaseGetCities}
\title{Возвращает список городов}
\usage{
databaseGetCities(country_id = "", region_id = "", q = "",
  need_all = "1", offset = "", count = "100")
}
\arguments{
\item{country_id}{Идентификатор страны, полученный в методе database.getCountries}

\item{region_id}{Идентификатор региона, города которого необходимо получить}

\item{q}{Строка поискового запроса}

\item{need_all}{1 – возвращать все города. 0 – возвращать только основные города}

\item{offset}{Отступ, необходимый для получения определенного подмножества городов}

\item{count}{Количество городов, которые необходимо вернуть}
}
\description{
Возвращает список городов
}
\examples{
\dontrun{
databaseGetCities(country_id=1, need_all=0)
}
}

