% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetCitiesById}
\alias{databaseGetCitiesById}
\title{Returns information about cities by their IDs}
\usage{
databaseGetCitiesById(city_ids = "")
}
\arguments{
\item{city_ids}{City IDs.}
}
\value{
Returns an array of city objects, each containing id and title fields.
}
\description{
Returns information about cities by their IDs
}
\examples{
databaseGetCitiesById('1,2')
}

