% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetCountries}
\alias{databaseGetCountries}
\title{Returns a list of countries}
\usage{
databaseGetCountries(need_all = "1", code = "", offset = "",
  count = "100")
}
\arguments{
\item{need_all}{1 — to return a full list of all countries, 0 — to return a list of countries near the current user's country (default).}

\item{code}{Country codes in ISO 3166-1 alpha-2 standard.}

\item{offset}{Offset needed to return a specific subset of countries.}

\item{count}{Number of countries to return.}
}
\value{
Returns an array of country objects, each containing the fields cid and title.
If need_all parameter is set, returns a full list of all countries.
If code parameter is set, returns a list of countries with their corresponding ISO 3166-1 alpha-2 codes.
}
\description{
Returns a list of countries
}
\examples{
databaseGetCountries(count=234)
}

