% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetFaculties}
\alias{databaseGetFaculties}
\title{Возвращает список факультетов}
\usage{
databaseGetFaculties(university_id = "", offset = "", count = "100")
}
\arguments{
\item{university_id}{Идентификатор университета, факультеты которого необходимо получить}

\item{offset}{Отступ, необходимый для получения определенного подмножества факультетов}

\item{count}{Количество факультетов которое необходимо получить}
}
\description{
Возвращает список факультетов
}
\examples{
\dontrun{
databaseGetFaculties(53)
}
}

