% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetRegions}
\alias{databaseGetRegions}
\title{Returns a list of regions}
\usage{
databaseGetRegions(country_id = "", q = "", offset = "", count = "100",
  v = getAPIVersion())
}
\arguments{
\item{country_id}{Country ID, received in database.getCountries method}

\item{q}{Search query}

\item{offset}{Offset needed to return specific subset of regions}

\item{count}{Number of regions to return}

\item{v}{Version of API}
}
\description{
Returns a list of regions
}
\examples{
\dontrun{
databaseGetRegions(229)
}
}

