% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{databaseGetSchoolClasses}
\alias{databaseGetSchoolClasses}
\title{Returns a list of available classes}
\usage{
databaseGetSchoolClasses(country_id = "", v = getAPIVersion())
}
\arguments{
\item{country_id}{Country ID}

\item{v}{Version of API}
}
\description{
Returns a list of available classes
}
\examples{
\dontrun{
databaseGetSchoolClasses(1)
}
}
