% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetSchools}
\alias{databaseGetSchools}
\title{Returns a list of schools}
\usage{
databaseGetSchools(q = "", city_id = "", offset = "", count = "100")
}
\arguments{
\item{q}{Search query.}

\item{city_id}{City ID.}

\item{offset}{Offset needed to return a specific subset of schools.}

\item{count}{Number of schools to return.}
}
\value{
Returns an array of school objects, each containing id and title fields.
}
\description{
Returns a list of schools
}
\examples{
databaseGetSchools(city_id = 2)
}

