% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{databaseGetStreetsById}
\alias{databaseGetStreetsById}
\title{Returns information about streets by their IDs}
\usage{
databaseGetStreetsById(street_ids = "", v = getAPIVersion())
}
\arguments{
\item{street_ids}{Street IDs}

\item{v}{Version of API}
}
\description{
Returns information about streets by their IDs
}
\examples{
\dontrun{
databaseGetStreetsById(1)
}
}
