% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetUniversities}
\alias{databaseGetUniversities}
\title{Возвращает список высших учебных заведений}
\usage{
databaseGetUniversities(q = "", country_id = "", city_id = "",
  offset = "", count = "100")
}
\arguments{
\item{q}{Строка поискового запроса}

\item{country_id}{Идентификатор страны, учебные заведения которой необходимо вернуть}

\item{city_id}{Идентификатор города, учебные заведения которого необходимо вернуть}

\item{offset}{Отступ, необходимый для получения определенного подмножества учебных заведений}

\item{count}{Количество учебных заведений, которое необходимо вернуть}
}
\description{
Возвращает список высших учебных заведений
}
\examples{
\dontrun{
databaseGetUniversities(city_id = '2')
}
}

