% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{db_drop_collection}
\alias{db_drop_collection}
\title{Drop collection}
\usage{
db_drop_collection(collection_name, collection_type)
}
\arguments{
\item{collection_name}{Collection name}

\item{collection_type}{Collection suffix}
}
\description{
Drop collection
}

