% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{db_get_connection}
\alias{db_get_connection}
\title{Get connection}
\usage{
db_get_connection(collection_name, collection_type)
}
\arguments{
\item{collection_name}{Collection name}

\item{collection_type}{Collection suffix}
}
\description{
Get connection
}

