% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{db_init}
\alias{db_init}
\title{Initialize database}
\usage{
db_init(db_name = "temp", verbose = FALSE)
}
\arguments{
\item{db_name}{Database name ('temp' by default)}

\item{verbose}{Emit some more output}
}
\description{
Initialize database
}

