% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{db_insert}
\alias{db_insert}
\title{Insert object into existing collection}
\usage{
db_insert(object, collection, suffix, db_name = db_getActive())
}
\arguments{
\item{object}{Object to insert}

\item{collection}{Collection name}

\item{suffix}{Collection suffix}

\item{db_name}{Database name}
}
\description{
Insert object into existing collection
}
