% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{db_load}
\alias{db_load}
\title{Load all collections from db for specified data base}
\usage{
db_load(db_name = db_getActive())
}
\arguments{
\item{db_name}{Database name}
}
\description{
Load all collections from db for specified data base
}

