% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{db_load_collection}
\alias{db_load_collection}
\title{Import collection from db to global env}
\usage{
db_load_collection(collection_name, collection_type)
}
\arguments{
\item{collection_name}{Collection name}

\item{collection_type}{Collection suffix}
}
\description{
Import collection from db to global env
}

