% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{db_save}
\alias{db_save}
\title{Save object to db}
\usage{
db_save(object, db_name = db_getActive(), collection, type)
}
\arguments{
\item{object}{Object to save}

\item{db_name}{Database name}

\item{collection}{Collection name}

\item{type}{Collection suffix}
}
\description{
Save object to db
}

