% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{db_update}
\alias{db_update}
\title{Update existing records}
\usage{
db_update(
  object,
  key,
  collection,
  suffix = "",
  db_name = db_getActive(),
  upsert = FALSE
)
}
\arguments{
\item{object}{Object to insert}

\item{key}{Key}

\item{collection}{Collection name}

\item{suffix}{Collection suffix}

\item{db_name}{Database name}

\item{upsert}{Insert a new document if no matching document exists}
}
\description{
Update existing records
}
