% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogr_wrapper.R
\name{demogr_wrapper}
\alias{demogr_wrapper}
\title{Run demogr on a set of variables and return a single output data frame}
\usage{
demogr_wrapper(idf, vois, group, numdec = 2, numdecp = 4, mediqr = FALSE)
}
\arguments{
\item{idf}{Input data frame}

\item{vois}{Outcome measure(s) to loop over}

\item{group}{Group variable to stratify the variable of interest by}

\item{numdec}{Number of decimals for mean/stdev  and mean/IQR variables}

\item{numdecp}{Number of decimals for p-values}

\item{mediqr}{If TRUE, then calculate median/IQR instead of mean/stdev}
}
\description{
This function is a wrapper script for demogr to loop over a number of
predcitor and outcome variables.
}
