% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2orgtable.R
\name{df2orgtable}
\alias{df2orgtable}
\title{Convert a data frame to an org table and print it}
\usage{
df2orgtable(
  df,
  colorizep = TRUE,
  thold = 0.05,
  rownamesout = FALSE,
  noNAs = TRUE,
  nozeros = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df}{Name of a data element in your working environment}

\item{colorizep}{When set to TRUE, automatically color columns ending in 'p'
and 'fdr' according the the threshold. If set to a list with one or two
arrays, the first array will be used as the p-value columns and the second
array will be used as fdr p-value columns.}

\item{thold}{Threshold for coloring}

\item{rownamesout}{If set to TRUE, print rownames in output table. If set to
a string, use this as the header of the rownames column and include it in the
output}

\item{noNAs}{Print cells with NA as empty}

\item{nozeros}{Print cells with 0 as empty}

\item{verbose}{Show messages}
}
\description{
Convert a data frame to an org table and print it
}
\examples{
df2orgtable(df,
            colorizep = list(c("col1", "col2"), c("col3", "col4")),
            thold = 0.05)
}
