% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{execute}
\alias{execute}
\title{Универсальный метод, который позволяет запускать последовательность других методов, сохраняя и фильтруя промежуточные результаты}
\usage{
execute(code)
}
\description{
Универсальный метод, который позволяет запускать последовательность других методов, сохраняя и фильтруя промежуточные результаты
}

