% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filterAttachments}
\alias{filterAttachments}
\title{Filtering attachments by type}
\usage{
filterAttachments(attachments, type)
}
\arguments{
\item{attachments}{List of attachments}

\item{type}{type field may have the following values:
\itemize {
\item \strong{photo} - photo from an album;
\item \strong{posted_photo} - photo uploaded directly from user's computer;
\item \strong{video} - video;
\item \strong{audio} - audio;
\item \strong{doc} - document;
\item \strong{graffiti} - graffiti;
\item \strong{url} - web page URL;
\item \strong{note} - note;
\item \strong{app} - image uploaded with a third party application;
\item \strong{poll} - poll;
\item \strong{page} - wiki page.
}}
}
\description{
Filtering attachments by type
}

