% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stop_words}
\alias{get_stop_words}
\title{Get stop words list for russian language}
\usage{
get_stop_words(stop_words = c())
}
\arguments{
\item{stop_words}{User defined stop words}
}
\description{
Get stop words list for russian language
}
