% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getCountryByCityId}
\alias{getCountryByCityId}
\title{Get country ID and title by given city ID}
\usage{
getCountryByCityId(city_id)
}
\arguments{
\item{city_id}{City ID}
}
\description{
Get country ID and title by given city ID
}

