% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFriends}
\alias{getFriends}
\title{Returns a list of user IDs or detailed information about a user's friends}
\usage{
getFriends(user_id = "", order = "", list_id = "", count = "",
  offset = "", fields = "", name_case = "", v = "5.33")
}
\arguments{
\item{user_id}{User ID. By default, the current user ID.}

\item{order}{Sort order: name — by name (enabled only if the fields parameter is used), hints — by rating, similar to how friends are sorted in My friends section.}

\item{list_id}{ID of the friend list returned by the getFriendsLists() method to be used as the source. This parameter is taken into account only when the user_id parameter is set to the current user ID.}

\item{count}{Number of friends to return.}

\item{offset}{Offset needed to return a specific subset of friends.}

\item{fields}{Profile fields to return.}

\item{name_case}{Case for declension of user name and surname:
\itemize{
\item \strong{nom} — nominative (default);
\item \strong{gen} — genitive;
\item \strong{dat} — dative;
\item \strong{acc} — accusative;
\item \strong{ins} — instrumental;
\item \strong{abl} — prepositional.
}}
}
\value{
If the fields parameter is not used, returns a list of IDs (id) of the user's friends.
If the fields parameter is used, returns an array of user objects, each containing a set of fields defined by the fields parameter.
The uid, first_name, last_name, lists, and online fields are always returned, regardless of the selected fields.
Returns not more than 5000 friends if filelds are used.
}
\description{
Returns a list of user IDs or detailed information about a user's friends
}
\examples{
friends_list <- getFriends(user_id='1', order='name', fields='bdate')
friends <- friends_list$items
}

