% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFriendsBy25}
\alias{getFriendsBy25}
\title{Получить список друзей для пользователей, указанных в векторе ids}
\usage{
getFriendsBy25(ids)
}
\arguments{
\item{ids}{Вектор со списком пользователей, для которых требуется получить список друзей}
}
\description{
Получить список друзей для пользователей, указанных в векторе ids
}

