% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{getFriendsFor}
\alias{getFriendsFor}
\title{Returns a list of friends IDs for the specified users}
\usage{
getFriendsFor(users_ids, v = getAPIVersion())
}
\arguments{
\item{users_ids}{User IDs}

\item{v}{Version of API}
}
\description{
Returns a list of friends IDs for the specified users
}
\examples{
\dontrun{
friends <- getFriendsFor(sample(x=seq(1:10000000), size=100, replace=FALSE))
users <- getUsersExecute(friends, fields = 'sex')
}
}

