% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{getGroupsById}
\alias{getGroupsById}
\title{Returns information about communities by their IDs}
\usage{
getGroupsById(group_ids = "", group_id = "", fields = "", v = getAPIVersion())
}
\arguments{
\item{group_ids}{IDs or screen names of communities}

\item{group_id}{ID or screen name of the community}

\item{fields}{Group fields to return}

\item{v}{Version of API}
}
\description{
Returns information about communities by their IDs
}
