% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{getGroupsForUsers}
\alias{getGroupsForUsers}
\title{Returns a list of the communities for the specified users}
\usage{
getGroupsForUsers(
  users,
  extended = "",
  filter = "",
  fields = "",
  progress_bar = FALSE,
  v = getAPIVersion()
)
}
\arguments{
\item{users}{A list of users}

\item{extended}{1 - to return complete information about a user's communities; 0 - to return a list of community IDs without any additional fields (default)}

\item{filter}{Types of communities to return: admin, editor, moder, groups, publics, events}

\item{fields}{List of additional fields to be returned}

\item{progress_bar}{Display progress bar}

\item{v}{Version of API}
}
\description{
Returns a list of the communities for the specified users
}
\examples{
\dontrun{
members <- getGroupsForUsers(c(me(), 123456), extended = 1, fields='city', progress_bar = TRUE)
}
}
