% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGroupsMembers}
\alias{getGroupsMembers}
\title{Возвращает список участников сообщества}
\usage{
getGroupsMembers(group_id = "", sort = "", offset = "", count = "",
  fields = "", filter = "", v = getAPIVersion())
}
\arguments{
\item{group_id}{Идентификатор или короткое имя сообщества}

\item{sort}{Сортировка, с которой необходимо вернуть список участников}

\item{offset}{Смещение, необходимое для выборки определенного подмножества участников}

\item{count}{Количество участников сообщества, информацию о которых необходимо получить (не более 1000)}

\item{fields}{Список дополнительных полей, которые необходимо вернуть}

\item{filter}{Доступны значения friends, unsure, managers}

\item{v}{Версия API}
}
\description{
Возвращает список участников сообщества
}
\examples{
\dontrun{
groups <- getGroupsMembers('1', fields='sex,bdate,city,country,photo_50,education,interests,music,movies,tv,books,games,about,quotes,personal')
}
}

