% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{getGroupsMembersExecute}
\alias{getGroupsMembersExecute}
\title{Returns a list of community members}
\usage{
getGroupsMembersExecute(group_id = "", fields = "", filter = "",
  flatten = FALSE, v = getAPIVersion())
}
\arguments{
\item{group_id}{ID or screen name of the community}

\item{fields}{List of additional fields to be returned}

\item{filter}{friends – only friends in this community will be returned; unsure – only those who pressed 'I may attend' will be returned (if it's an event)}

\item{flatten}{Automatically flatten nested data frames into a single non-nested data frame}

\item{v}{Version of API}
}
\description{
Returns a list of community members
}

