% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMutual}
\alias{getMutual}
\title{Returns a list of user IDs of the mutual friends of two users.}
\usage{
getMutual(source_id = "", target_uid = "", target_uids = "", order = "",
  count = "", offset = "")
}
\arguments{
\item{source_id}{ID of the user whose friends will be checked against the friends of the user specified in target_uid.}

\item{target_uid}{ID of the user whose friends will be checked against the friends of the user specified in source_uid.}

\item{target_uids}{List comma-separated positive numbers.}

\item{order}{Sort order: random — random order.}

\item{count}{Number of mutual friends to return.}

\item{offset}{Offset needed to return a specific subset of mutual friends.}
}
\value{
Returns a list of IDs (id) of the mutual friends of the source_uid and target_uid users.
}
\description{
Returns a list of user IDs of the mutual friends of two users.
}
\examples{
mutual_friends <- getMutual(target_uid='1')
}

