% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{getPaths}
\alias{getPaths}
\title{Returns a list of paths between two users}
\usage{
getPaths(source_id, target_id, are_friends = FALSE, max_depth = 5)
}
\arguments{
\item{source_id}{Source ID}

\item{target_id}{Target ID}

\item{are_friends}{By default is FALSE}

\item{max_depth}{Maximum depth}
}
\description{
Returns a list of paths between two users
}
