% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getURLs}
\alias{getURLs}
\title{Extract URLs from messages}
\usage{
getURLs(messages, message_body = FALSE)
}
\arguments{
\item{messages}{Array of messages}

\item{message_body}{Add message body to URLs}
}
\description{
Extract URLs from messages
}
