% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getUsers}
\alias{getUsers}
\title{Возвращает расширенную информацию о пользователях}
\usage{
getUsers(user_ids = "", fields = "", name_case = "")
}
\arguments{
\item{user_ids}{Перечисленные через запятую идентификаторы пользователей или их короткие имена}

\item{fields}{Список дополнительных полей профилей, которые необходимо вернуть}

\item{name_case}{Падеж для склонения имени и фамилии пользователя}
}
\description{
Возвращает расширенную информацию о пользователях
}
\examples{
user <- getUsers('1', fields='sex,bdate,city,country,photo_50,education,interests,music,movies,tv,books,games,about,quotes,personal')
}

