% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{getUsersExecute}
\alias{getUsersExecute}
\title{Returns detailed information on arbitrary number of users}
\usage{
getUsersExecute(users_ids = "", fields = "", name_case = "",
  v = getAPIVersion())
}
\arguments{
\item{fields}{Profile fields to return}

\item{name_case}{Case for declension of user name and surname}

\item{v}{Version of API}

\item{user_ids}{User IDs or screen names (screen_name). By default, current user ID}
}
\description{
Returns detailed information on arbitrary number of users
}
\examples{
\dontrun{
users <- getUsersExecute(sample(x=seq(1:10000000), size=10000, replace=FALSE), fields='sex,bdate,city')
}
}

