% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWall}
\alias{getWall}
\title{Возвращает список записей со стены пользователя или сообщества}
\usage{
getWall(owner_id = "", domain = "", offset = "", count = "",
  filter = "owner", extended = "", v = "5.28")
}
\arguments{
\item{owner_id}{Идентификатор пользователя или сообщества}

\item{domain}{Короткий адрес пользователя или сообщества}

\item{offset}{Смещение, необходимое для выборки определенного подмножества записей}

\item{count}{Количество записей, которое необходимо получить (но не более 100)}

\item{filter}{Определяет, какие типы записей на стене необходимо получить}

\item{extended}{Если установлен в 1, то будут возвращены три массива wall, profiles и groups
getWall()}
}
\description{
Возвращает список записей со стены пользователя или сообщества
}

