% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{h1}
\alias{h1}
\title{Print a nice header}
\usage{
h1(my_text)
}
\arguments{
\item{my_text}{Text that you want to use for a header
Header 1}
}
\description{
This function prints nice headers around a string input.
}
