% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html}
\alias{html}
\title{Display data in your browser}
\usage{
html(df, type = NULL)
}
\arguments{
\item{df}{Name of a data element in your working environment}

\item{type}{Simple (s) or beautified (b) output}
}
\description{
This function loads a data element in your browser
for quick display.
}
