% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_column.R
\name{insert_column}
\alias{insert_column}
\title{Add a column inside a dataframe}
\usage{
insert_column(df, position, colname, values)
}
\arguments{
\item{df}{Input data frame}

\item{position}{Column (either a column number or column name AFTER which you want to insert your new column}

\item{colname}{Name of the new column}

\item{values}{Values of the new column. Either a single value (that will be repeated for each row) or an array of length nrow(df) are accepted.}
}
\description{
Add a column inside a dataframe
}
