% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install and/or update a package without bugging for selecting a repo}
\usage{
install(my_package)
}
\arguments{
\item{my_package}{Name of the package to be installed}
}
\description{
Install and/or update a package without bugging for selecting a repo
}
