% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_my_packages.R
\name{install_my_packages}
\alias{install_my_packages}
\title{Install and/or update all packages I use frequently}
\usage{
install_my_packages()
}
\description{
Install and/or update all packages I use frequently
}
