% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libv.R
\name{libv}
\alias{libv}
\title{Display data in Libreoffice from R}
\usage{
libv(x, filepath = NULL)
}
\arguments{
\item{x}{Name of a data element in your working environment}

\item{filepath}{Path to data file}
}
\description{
This function loads a dataframe in Libreoffice
}
