% Generated by roxygen2 (4.0.2): do not edit by hand
\name{likesGetList}
\alias{likesGetList}
\title{Получает список идентификаторов пользователей, которые добавили заданный объект в свой список "Мне нравится"}
\usage{
likesGetList(type = "", owner_id = "", item_id = "", page_url = "",
  filter = "", friends_only = "0", extended = "", offset = "",
  count = "100")
}
\arguments{
\item{type}{Тип объекта}

\item{owner_id}{Идентификатор владельца Like-объекта: id пользователя, id сообщества (со знаком «минус») или id приложения}

\item{item_id}{Идентификатор Like-объекта}

\item{page_url}{url страницы, на которой установлен виджет "Мне нравится"}

\item{filter}{Указывает, следует ли вернуть всех пользователей, добавивших объект в список "Мне нравится" или только тех, которые рассказали о нем друзьям}

\item{friends_only}{Указывает, необходимо ли возвращать только пользователей, которые являются друзьями текущего пользователя}

\item{extended}{1 — возвращать расширенную информацию о пользователях и сообществах из списка поставивших отметку «Мне нравится» или сделавших репост}

\item{offset}{Смещение, относительно начала списка, для выборки определенного подмножества}

\item{count}{Количество возвращаемых идентификаторов пользователей}
}
\description{
Получает список идентификаторов пользователей, которые добавили заданный объект в свой список "Мне нравится"
}

