% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{messagesGet}
\alias{messagesGet}
\title{Returns a list of the current user's incoming or outgoing private messages}
\usage{
messagesGet(out = "", offset = "", count = "", time_offset = "",
  filters = "", preview_length = "", last_message_id = "",
  v = getAPIVersion())
}
\arguments{
\item{out}{1 — to return outgoing messages; 0 — to return incoming messages (default)}

\item{offset}{Offset needed to return a specific subset of messages}

\item{count}{Number of messages to return}

\item{time_offset}{Maximum time since a message was sent, in seconds. To return messages without a time limitation, set as 0}

\item{filters}{Filter to apply:  1 — unread only; 2 — not from the chat; 4 — messages from friends}

\item{preview_length}{Number of characters after which to truncate a previewed message. To preview the full message, specify 0}

\item{last_message_id}{ID of the message received before the message that will be returned last}

\item{v}{Version of API}
}
\description{
Returns a list of the current user's incoming or outgoing private messages
}

