% Generated by roxygen2 (4.0.2): do not edit by hand
\name{messagesGetHistory}
\alias{messagesGetHistory}
\title{Возвращает историю сообщений для указанного пользователя}
\usage{
messagesGetHistory(offset = "", count = "", user_id = "", chat_id = "",
  start_message_id = "", rev = "", v = getAPIVersion())
}
\arguments{
\item{offset}{Смещение, необходимое для выборки определенного подмножества сообщений}

\item{count}{Количество сообщений, которое необходимо получить (но не более 200)}

\item{user_id}{Идентификатор пользователя, историю переписки с которым необходимо вернуть}

\item{chat_id}{Идентификатор диалога, историю сообщений которого необходимо получить}

\item{start_message_id}{Если установлен в 1, то сообщения будут возвращены в хронологическом порядке}

\item{rev}{1 – возвращать сообщения в хронологическом порядке}

\item{v}{Версия API}
}
\description{
Возвращает историю сообщений для указанного пользователя
}

