% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{messagesGetHistoryAll}
\alias{messagesGetHistoryAll}
\title{Returns all message history for the specified user or group chat}
\usage{
messagesGetHistoryAll(user_id = "", peer_id = "", rev = 0,
  v = getAPIVersion())
}
\arguments{
\item{user_id}{ID of the user whose message history you want to return}
}
\description{
Returns all message history for the specified user or group chat
}

