% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{messagesGetHistoryExecute}
\alias{messagesGetHistoryExecute}
\title{Returns message history for the specified user or group chat}
\usage{
messagesGetHistoryExecute(offset = 0, count = 0, user_id = "",
  peer_id = "", start_message_id = "", rev = 0, progress_bar = FALSE,
  v = getAPIVersion())
}
\arguments{
\item{offset}{Offset needed to return a specific subset of messages}

\item{count}{Number of messages to return (0 for all history)}

\item{user_id}{ID of the user whose message history you want to return}

\item{peer_id}{Destination ID (user ID, group ID or chat ID)}

\item{start_message_id}{Starting message ID from which to return history}

\item{rev}{Sort order: 1 - return messages in chronological order; 0 - return messages in reverse chronological order}

\item{progress_bar}{Display progress bar}

\item{v}{Version of API}
}
\description{
Returns message history for the specified user or group chat
}
