% Generated by roxygen2 (4.0.2): do not edit by hand
\name{messagesSplitByDate}
\alias{messagesSplitByDate}
\title{Разбить сообщения по дням, неделям, месецам}
\usage{
messagesSplitByDate(messages, format = "\%y-\%m-\%d")
}
\arguments{
\item{messages}{Сообщения, полученные с помощью функций messagesGet()}

\item{format}{Формат даты, по которому будет происходить группировка сообщений}
}
\description{
Разбить сообщения по дням, неделям, месецам
}

