% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm_fdr.R
\name{my_lm_fdr}
\alias{my_lm_fdr}
\title{Apply FDR correction to my_lm_lm output}
\usage{
my_lm_fdr(idf)
}
\arguments{
\item{idf}{Input data frame}
}
\description{
This function takes output from my_lm_lm and grabs all the values in the
columns ending in _fdr. my_lm_lm stores full length (not truncated) raw
p-values there. It creates one array of these p-values and then applies FDR
correction to it. After that, it stores the corrected p-values back into
these columns.
}
