% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm_lm.R
\name{my_lm_lm}
\alias{my_lm_lm}
\title{Print summary statistics for a linear model}
\usage{
my_lm_lm(df, outcome, predictor, covariates = NULL, verbose = FALSE)
}
\arguments{
\item{df}{Input data frame}

\item{outcome}{Outcome measure}

\item{predictor}{Predictor}

\item{covariates}{Single covariate or array of covariates}

\item{verbose}{Set to TRUE if you want to see the full lm output}
}
\description{
This function prints summary statistics for a linear model such as the
estimate, std.err, p-value, and effect size for a predictor. It will
automatically detect if the predictor variable is numeric or a factor
variable. If it is a factor variable, then estimates etc. will be provided
for all levels of the facor variable (except the first level, that will be
the reference categorie)
}
