% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsfeed.R
\name{newsfeedSearch}
\alias{newsfeedSearch}
\title{Returns search results by statuses}
\usage{
newsfeedSearch(
  q = "",
  extended = "",
  count = "",
  latitude = "",
  longitude = "",
  start_time = "",
  end_time = "",
  start_from = "",
  fields = "",
  v = getAPIVersion()
)
}
\arguments{
\item{q}{Search query string (e.g., New Year).}

\item{extended}{1 — to return additional information about the user or community that placed the post}

\item{count}{Number of posts to return}

\item{latitude}{Geographical latitude point (in degrees, -90 to 90) within which to search}

\item{longitude}{Geographical longitude point (in degrees, -180 to 180) within which to search}

\item{start_time}{Earliest timestamp (in Unix time) of a news item to return. By default, 24 hours ago}

\item{end_time}{Latest timestamp (in Unix time) of a news item to return. By default, the current time}

\item{start_from}{String, accessible for versions from 5.13}

\item{fields}{Additional fields of profiles and communities to return}

\item{v}{Version of API}
}
\description{
Returns search results by statuses
}
