% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Display Normality and Normalize Data}
\usage{
normalize(
  df,
  inputs,
  transform = FALSE,
  truncsd = FALSE,
  plotdir = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df}{Input data frame}

\item{inputs}{List of input variables that are being processed}

\item{transform}{If TRUE, apply log transformation}

\item{truncsd}{If set to a value, truncate the log transformed data to mean
+/- truncsd}

\item{plotdir}{If set to a path, it will output the normality plots there}

\item{verbose}{If TRUE, display verbose output}
}
\description{
This function provides an overview of the normality of data and allows for
normalizing data that is not normally distributed
}
