% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wall.R
\name{postGetComments}
\alias{postGetComments}
\title{Returns a list of comments on a post on a user wall or community wall}
\usage{
postGetComments(owner_id = "", post_id = "", need_likes = 1,
  start_comment_id = "", offset = 0, count = 10, sort = "",
  preview_length = 0, extended = "", progress_bar = FALSE,
  v = getAPIVersion())
}
\arguments{
\item{owner_id}{User ID or community ID. Use a negative value to designate a community ID.}

\item{post_id}{Post ID.}

\item{need_likes}{1 - to return the likes field (default), 0 - not to return the likes field.}

\item{start_comment_id}{Positive number}

\item{offset}{Offset needed to return a specific subset of comments.}

\item{count}{Number of comments to return.}

\item{sort}{Sort order: asc - chronological, desc - reverse chronological.}

\item{preview_length}{Number of characters at which to truncate comments when previewed. Specify 0 (default) if you do not want to truncate comments.}

\item{extended}{Flag, either 1 or 0.}

\item{progress_bar}{Display progress bar}

\item{v}{Version of API}
}
\description{
Returns a list of comments on a post on a user wall or community wall
}
