% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{profile_fields}
\alias{profile_fields}
\title{Helper function for working with profile fields}
\usage{
profile_fields(fields = "")
}
\arguments{
\item{fields}{Profile fields to return}
}
\description{
Helper function for working with profile fields
}
\examples{
\dontrun{
# get list of all fields
fields <- profile_fields('all')

# get list of all fields except specified
fields <- profile_fields('all - photo_50,photo_100,photo_200')

# get only specified fields
fields <- profile_fields('sex,bdate')
}
}

