% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queryBuilder}
\alias{queryBuilder}
\title{Возвращает строку запроса, созданную из переданных аргументов}
\usage{
queryBuilder(method_name, ...)
}
\arguments{
\item{method_name}{Имя метода}
}
\description{
Возвращает строку запроса, созданную из переданных аргументов
}

