% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPCA.R
\name{quick_pca}
\alias{quick_pca}
\title{Quick Principal component Analysis}
\usage{
quick_pca(input_data, rotation, out_path, out_name)
}
\arguments{
\item{input_data}{Data frame for PCA}

\item{rotation}{Type of rotation (varimax or oblimin)}

\item{out_path}{Path for storing the scree plot}

\item{out_name}{File name for the scree plot}
}
\description{
Quick Principal component Analysis
}
