% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_gefx_exporter.R
\name{saveAsGEXF}
\alias{saveAsGEXF}
\title{Converts the given igraph object to GEXF format and saves it at the given filepath location}
\usage{
saveAsGEXF(g, filepath = "converted_graph.gexf")
}
\arguments{
\item{g}{Input igraph object to be converted to gexf format}

\item{filepath}{File location where the output gexf file should be saved}
}
\description{
Converts the given igraph object to GEXF format and saves it at the given filepath location
}
\author{
Gopalakrishna Palem, \email{Gopalakrishna.Palem@Yahoo.com}
}

