% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{try_handle_error}
\alias{try_handle_error}
\title{Check response for errors}
\usage{
try_handle_error(response)
}
\arguments{
\item{response}{httr response object}
}
\description{
Check response for errors
}
