% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{try_handle_network_error}
\alias{try_handle_network_error}
\title{Try to handle network error}
\usage{
try_handle_network_error(error)
}
\arguments{
\item{error}{Error}
}
\description{
Try to handle network error
}
