% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{usersGetSubscriptions}
\alias{usersGetSubscriptions}
\title{Returns a list of IDs of users and communities followed by the user}
\usage{
usersGetSubscriptions(user_id = "", extended = "1", offset = 0,
  count = 0, fields = "", flatten = FALSE, progress_bar = FALSE,
  v = getAPIVersion())
}
\arguments{
\item{user_id}{User ID}

\item{extended}{1 - to return a combined list of users and communities, 0 - to return separate lists of users and communities}

\item{offset}{Offset needed to return a specific subset of subscriptions}

\item{count}{Number of users and communities to return}

\item{fields}{Profile fields to return}

\item{flatten}{Automatically flatten nested data frames into a single non-nested data frame}

\item{progress_bar}{Display progress bar}

\item{v}{Version of API}
}
\description{
Returns a list of IDs of users and communities followed by the user
}
\examples{
\dontrun{
my_subscriptions <- usersGetSubscriptions(me())
}
}
