% Generated by roxygen2 (4.0.2): do not edit by hand
\name{usersSearch}
\alias{usersSearch}
\title{Возвращает список пользователей в соответствии с заданным критерием поиска}
\usage{
usersSearch(q = "", sort = "", offset = "", count = "20", fields = "",
  city = "", country = "", hometown = "", university_country = "",
  university = "", university_year = "", university_faculty = "",
  university_chair = "", sex = "", status = "", age_from = "",
  age_to = "", birth_day = "", birth_month = "", birth_year = "",
  online = "", has_photo = "", school_country = "", school_city = "",
  school_class = "", school = "", school_year = "", religion = "",
  interests = "", company = "", position = "", group_id = "",
  v = getAPIVersion())
}
\arguments{
\item{q}{строка поискового запроса. Например, Вася Бабич}

\item{sort}{сортировка результатов: 1 - по дате регистрации, 0 - по популярности}

\item{count}{количество возвращаемых пользователей}

\item{fields}{список дополнительных полей, которые необходимо вернуть}

\item{city}{идентификатор города}

\item{country}{идентификатор страны}

\item{hometown}{название города строкой}

\item{university_country}{идентификатор страны, в которой пользователи закончили ВУЗ}

\item{university}{идентификатор ВУЗа}

\item{university_year}{год окончания ВУЗа}

\item{university_faculty}{идентификатор факультета}

\item{university_chair}{идентификатор кафедры}

\item{sex}{пол, 1 — женщина, 2 — мужчина, 0 (по умолчанию) — любой}

\item{status}{семейное положение: 1 — Не женат, 2 — Встречается, 3 — Помолвлен, 4 — Женат, 7 — Влюблён, 5 — Всё сложно, 6 — В активном поиске}

\item{age_from}{начиная с какого возраста}

\item{age_to}{до какого возраста}

\item{birth_day}{день рождения}

\item{birth_month}{месяц рождения}

\item{birth_year}{год рождения}

\item{online}{1 — только в сети, 0 — все пользователи}

\item{has_photo}{1 — только с фотографией, 0 — все пользователи}

\item{school_country}{идентификатор страны, в которой пользователи закончили школу}

\item{school_city}{идентификатор города, в котором пользователи закончили школу}

\item{school_class}{положительное число}

\item{school}{идентификатор школы, которую закончили пользователи}

\item{school_year}{год окончания школы}

\item{religion}{религиозные взгляды}

\item{interests}{интересы}

\item{company}{название компании, в которой работают пользователи}

\item{position}{название должности}

\item{group_id}{идентификатор группы, среди пользователей которой необходимо проводить поиск}

\item{v}{Версия API}

\item{from_list}{Разделы среди которых нужно осуществить поиск, перечисленные через запятую: friends – искать среди друзей, subscribes – искать среди друзей и подписок пользователя}
}
\description{
Возвращает список пользователей в соответствии с заданным критерием поиска
}

