% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{vk_stop}
\alias{vk_stop}
\title{Custom error}
\usage{
vk_stop(message = "", call = sys.call(), error_code = "")
}
\arguments{
\item{message}{Error message}

\item{call}{Call expression}

\item{error_code}{Error code}
}
\description{
Custom error
}

