% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{vkOAuthWeb}
\alias{vkOAuthWeb}
\title{Client authorization (for web application)}
\usage{
vkOAuthWeb(app_name, client_id, client_secret)
}
\arguments{
\item{app_name}{Application name}

\item{client_id}{Application ID}

\item{client_secret}{Application secret key}
}
\description{
Client authorization (for web application)
}

