% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vkR.R
\docType{package}
\name{vkR}
\alias{vkR}
\title{Access to VK API via R}
\description{
This package provides a series of functions that allow R users
to access VK's API (\url{https://vk.com/dev/methods}) to get
information about users, messages, groups, posts and likes.
}
\details{
VK (\url{https://vk.com/}) is the largest European online social
networking service, based in Russia. It is available in several
languages, and is especially popular among Russian-speaking users.
VK allows users to message each other publicly or privately, to
create groups, public pages and events, share and tag images,
audio and video, and to play browser-based games [1].
}
\references{
[1] \url{https://en.wikipedia.org/wiki/VK_(social_networking)}
}
\seealso{
\code{\link{vkOAuth}},
\code{\link{getUsersExecute}}, \code{\link{getWallExecute}},
\code{\link{getFriends}}, \code{\link{getFriendsFor}},
\code{\link{getGroupsForUsers}}, \code{\link{getGroupsMembersExecute}},
\code{\link{likesGetListForObjects}}, \code{\link{messagesGetHistoryExecute}},
\code{\link{getArbitraryNetwork}}, \code{\link{getStatus}}
}
\author{
Dmitriy Sorokin \email{dementiy@yandex.ru}
}
