% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wallGetComments}
\alias{wallGetComments}
\title{Возвращает список комментариев к записи на стене}
\usage{
wallGetComments(owner_id = "", post_id = "", need_likes = "",
  offset = "", count = "10", sort = "", preview_length = "0",
  extended = "")
}
\arguments{
\item{owner_id}{Идентификатор владельца страницы}

\item{post_id}{Идентификатор записи на стене}

\item{need_likes}{1 — возвращать информацию о лайках}

\item{offset}{Сдвиг, необходимый для получения конкретной выборки результатов}

\item{count}{Число комментариев, которые необходимо получить}

\item{sort}{Порядок сортировки комментариев (asc — от старых к новым, desc - от новых к старым)}

\item{preview_length}{Количество символов, по которому нужно обрезать текст комментария}

\item{extended}{1 - комментарии в ответе будут возвращены в виде пронумерованных объектов, дополнительно будут возвращены списки объектов profiles, groups}
}
\description{
Возвращает список комментариев к записи на стене
}

