% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wallGetReposts}
\alias{wallGetReposts}
\title{Позволяет получать список репостов заданной записи}
\usage{
wallGetReposts(owner_id = "", post_id = "", offset = "", count = "20")
}
\arguments{
\item{owner_id}{Идентификатор пользователя или сообщества, на стене которого находится запись}

\item{post_id}{Идентификатор записи на стене}

\item{offset}{Смещение, необходимое для выборки определенного подмножества записей}

\item{count}{Количество записей, которое необходимо получить}
}
\description{
Позволяет получать список репостов заданной записи
}

