% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wallSearch}
\alias{wallSearch}
\title{Allows to search posts on user or community walls}
\usage{
wallSearch(owner_id = "", domain = "", query = "", owners_only = "",
  count = "20", offset = "0", extended = "")
}
\arguments{
\item{owner_id}{User or community id. Remember that for a community owner_id must be negative.}

\item{domain}{User or community screen name.}

\item{query}{Search query string.}

\item{owners_only}{1 – returns only page owner's posts.}

\item{count}{Count of posts to return.}

\item{offset}{Results offset.}

\item{extended}{Show extended post info.}
}
\value{
If executed successfully, returns a list of post objects.
}
\description{
Allows to search posts on user or community walls
}

