% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_lm_wrapper.R
\name{batch_lm_wrapper}
\alias{batch_lm_wrapper}
\title{Run batch_lm and subsquently batch_lm_fdr on a number of predictor and
outcome variables.}
\usage{
batch_lm_wrapper(
  df,
  outcomes,
  predictors,
  covariates = NULL,
  numdec = 2,
  verbose = FALSE
)
}
\arguments{
\item{df}{Input data frame}

\item{outcomes}{Outcome measure(s) to loop over}

\item{predictors}{Predictor(s) to loop over}

\item{covariates}{Single covariate or array of covariates}

\item{numdec}{Number of decimals for the beta and std statistics}

\item{verbose}{Set to TRUE if you want to see the full lm output}
}
\description{
This function is a wrapper script for batch_lm to loop over a number of
predcitor and outcome variables, and to subsequently apply FDR correction on
the output using batch_lm_fdr.
}
