% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_esize.R
\name{convert_esize}
\alias{convert_esize}
\title{Convert effect-sizes}
\usage{
convert_esize(my_e, my_t)
}
\arguments{
\item{Value}{of an effect size}

\item{Type}{of effect size (h, f, d)}
}
\description{
This function converts between the following effect sizes:
Cohen's d, Cohen's f, and Eta-squared
The formulas that are used are from:
Cohen J., Statistical Power Analysis for the Behavioral
Sciences 2nd Ed (1988)
}
\details{
\enumerate{
\item Cohen's f to Cohen's d
d = 2f
(Page 276)
\item Cohen's f to eta-squared
eta-squared = f^2 / (1 + f^2)
(Page 281)
}

Interpretation of the effec sizes is as follows:
(https://en.wikipedia.org/wiki/Effect_size)
(http://core.ecu.edu/psyc/wuenschk/docs30/EffectSizeConventions.pdf)
(https://imaging.mrc-cbu.cam.ac.uk/statswiki/FAQ/effectSize)

Cohen's d:
small: 0.2
medium: 0.5
large: 0.8

Cohen's f:
small: 0.10
medium: 0.25
large: 0.40

eta-squared:
small: 0.01
medium: 0.06
large: 0.14
}
