% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetChairs}
\alias{databaseGetChairs}
\title{Возвращает список кафедр университета по указанному факультету}
\usage{
databaseGetChairs(faculty_id = "", offset = "", count = "100")
}
\arguments{
\item{faculty_id}{Идентификатор факультета, кафедры которого необходимо получить}

\item{offset}{Отступ, необходимый для получения определенного подмножества кафедр}

\item{count}{Количество кафедр которое необходимо получить}
}
\description{
Возвращает список кафедр университета по указанному факультету
}
\examples{
\dontrun{
databaseGetChairs(206)
}
}

