% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{databaseGetChairs}
\alias{databaseGetChairs}
\title{Returns list of chairs on a specified faculty}
\usage{
databaseGetChairs(faculty_id = "", offset = "", count = "100",
  v = getAPIVersion())
}
\arguments{
\item{faculty_id}{ID of the faculty to get chairs from}

\item{offset}{Offset required to get a certain subset of chairs}

\item{count}{Amount of chairs to get}

\item{v}{Version of API}
}
\description{
Returns list of chairs on a specified faculty
}
\examples{
\dontrun{
databaseGetChairs(206)
}
}

