% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{databaseGetCities}
\alias{databaseGetCities}
\title{Returns a list of cities}
\usage{
databaseGetCities(
  country_id = "",
  region_id = "",
  q = "",
  need_all = "1",
  offset = "",
  count = "100",
  v = getAPIVersion()
)
}
\arguments{
\item{country_id}{Country ID}

\item{region_id}{Region ID}

\item{q}{Search query}

\item{need_all}{1 - to return all cities in the country; 0 - to return major cities in the country (default)}

\item{offset}{Offset needed to return a specific subset of cities}

\item{count}{Number of cities to return}

\item{v}{Version of API}
}
\description{
Returns a list of cities
}
\examples{
\dontrun{
databaseGetCities(country_id=1, need_all=0)
}
}
