% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetCities}
\alias{databaseGetCities}
\title{Returns a list of cities}
\usage{
databaseGetCities(country_id = "", region_id = "", q = "",
  need_all = "1", offset = "", count = "100")
}
\arguments{
\item{country_id}{Country ID.}

\item{region_id}{Region ID.}

\item{q}{Search query.}

\item{need_all}{1 — to return all cities in the country, 0 — to return major cities in the country (default).}

\item{offset}{Offset needed to return a specific subset of cities.}

\item{count}{Number of cities to return.}
}
\value{
Returns an array of city objects, each containing cid and title fields. Object contains additional area and region fields.
}
\description{
Returns a list of cities
}
\examples{
databaseGetCities(country_id=1, need_all=0)
}

