% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetCitiesById}
\alias{databaseGetCitiesById}
\title{Возвращает информацию о городах по их идентификаторам}
\usage{
databaseGetCitiesById(city_ids = "")
}
\arguments{
\item{city_ids}{Идентификаторы городов}
}
\description{
Возвращает информацию о городах по их идентификаторам
}
\examples{
\dontrun{
databaseGetCitiesById('1,2')
}
}

