% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetCitiesById}
\alias{databaseGetCitiesById}
\title{Returns information about cities by their IDs}
\usage{
databaseGetCitiesById(city_ids = "", v = getAPIVersion())
}
\arguments{
\item{city_ids}{City IDs}

\item{v}{Version of API}
}
\description{
Returns information about cities by their IDs
}
\examples{
\dontrun{
databaseGetCitiesById('1,2')
}
}

