% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetCountries}
\alias{databaseGetCountries}
\title{Возвращает список стран}
\usage{
databaseGetCountries(need_all = "1", code = "", offset = "",
  count = "100")
}
\arguments{
\item{need_all}{1 - вернуть список всех стран. Флаг, может принимать значения 1 или 0}

\item{code}{Перечисленные через запятую двухбуквенные коды стран в стандарте ISO 3166-1 alpha-2, для которых необходимо выдать информацию}

\item{offset}{Отступ, необходимый для выбора определенного подмножества стран}

\item{count}{Количество стран, которое необходимо вернуть}
}
\description{
Возвращает список стран
}
\examples{
databaseGetCountries(count=234)
}

