% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{databaseGetCountries}
\alias{databaseGetCountries}
\title{Returns a list of countries}
\usage{
databaseGetCountries(need_all = "1", code = "", offset = "",
  count = "100", v = getAPIVersion())
}
\arguments{
\item{need_all}{1 - to return a full list of all countries; 0 - to return a list of countries near the current user's country}

\item{code}{Country codes in ISO 3166-1 alpha-2 standard}

\item{offset}{Offset needed to return a specific subset of countries}

\item{count}{Number of countries to return}

\item{v}{Version of API}
}
\description{
Returns a list of countries
}
\examples{
\dontrun{
databaseGetCountries(count=234)
}
}
