% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetCountriesById}
\alias{databaseGetCountriesById}
\title{Возвращает информацию о странах по их идентификаторам}
\usage{
databaseGetCountriesById(country_ids)
}
\arguments{
\item{country_ids}{Идентификаторы стран}
}
\description{
Возвращает информацию о странах по их идентификаторам
}
\examples{
\dontrun{
databaseGetCountriesById('1,2,3,4')
}
}

