% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{databaseGetCountriesById}
\alias{databaseGetCountriesById}
\title{Returns information about countries by their IDs}
\usage{
databaseGetCountriesById(country_ids, v = getAPIVersion())
}
\arguments{
\item{country_ids}{Country IDs}

\item{v}{Version of API}
}
\description{
Returns information about countries by their IDs
}
\examples{
\dontrun{
databaseGetCountriesById('1,2,3,4')
}
}
