% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetCountriesById}
\alias{databaseGetCountriesById}
\title{Returns information about countries by their IDs}
\usage{
databaseGetCountriesById(country_ids)
}
\arguments{
\item{country_ids}{Country IDs.}
}
\value{
Returns an array of country objects, each containing cid and name fields.
}
\description{
Returns information about countries by their IDs
}
\examples{
databaseGetCountriesById('1,2,3,4')
}

