% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{databaseGetFaculties}
\alias{databaseGetFaculties}
\title{Returns a list of faculties (i.e., university departments)}
\usage{
databaseGetFaculties(university_id = "", offset = "", count = "100",
  v = getAPIVersion())
}
\arguments{
\item{university_id}{University ID}

\item{offset}{Offset needed to return a specific subset of faculties}

\item{count}{Number of faculties to return}

\item{v}{Version of API}
}
\description{
Returns a list of faculties (i.e., university departments)
}
\examples{
\dontrun{
databaseGetFaculties(53)
}
}
