% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetFaculties}
\alias{databaseGetFaculties}
\title{Returns a list of faculties (i.e., university departments)}
\usage{
databaseGetFaculties(university_id = "", offset = "", count = "100")
}
\arguments{
\item{university_id}{University ID.}

\item{offset}{Offset needed to return a specific subset of faculties.}

\item{count}{Number of faculties to return.}
}
\value{
Returns a list of faculty objects with id and title fields.
}
\description{
Returns a list of faculties (i.e., university departments)
}
\examples{
databaseGetFaculties(53)
}

