% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetRegions}
\alias{databaseGetRegions}
\title{Возвращает список регионов}
\usage{
databaseGetRegions(country_id = "", q = "", offset = "", count = "100")
}
\arguments{
\item{country_id}{Идентификатор страны, полученный в методе database.getCountries}

\item{q}{Строка поискового запроса}

\item{offset}{Отступ, необходимый для выбора определенного подмножества регионов}

\item{count}{Количество регионов, которое необходимо вернуть}
}
\description{
Возвращает список регионов
}
\examples{
databaseGetRegions(229)
}

