% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetSchoolClasses}
\alias{databaseGetSchoolClasses}
\title{Возвращает список классов, характерных для школ определенной страны (!!!Warning!!!)}
\usage{
databaseGetSchoolClasses(country_id = "")
}
\arguments{
\item{country_id}{Идентификатор страны, доступные классы в которой необходимо вернуть}
}
\description{
Возвращает список классов, характерных для школ определенной страны (!!!Warning!!!)
}
\examples{
databaseGetSchoolClasses(1)
}

