% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetSchools}
\alias{databaseGetSchools}
\title{Возвращает список школ}
\usage{
databaseGetSchools(q = "", city_id = "", offset = "", count = "100")
}
\arguments{
\item{q}{Строка поискового запроса}

\item{city_id}{Идентификатор города, школы которого необходимо вернуть}

\item{offset}{Отступ, необходимый для получения определенного подмножества школ}

\item{count}{Количество школ, которое необходимо вернуть}
}
\description{
Возвращает список школ
}
\examples{
\dontrun{
databaseGetSchools(city_id = 2)
}
}

