% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{databaseGetSchools}
\alias{databaseGetSchools}
\title{Returns a list of schools}
\usage{
databaseGetSchools(q = "", city_id = "", offset = "", count = "100",
  v = getAPIVersion())
}
\arguments{
\item{q}{Search query}

\item{city_id}{City ID}

\item{offset}{Offset needed to return a specific subset of schools}

\item{count}{Number of schools to return}

\item{v}{Version of API}
}
\description{
Returns a list of schools
}
\examples{
\dontrun{
databaseGetSchools(city_id = 2)
}
}

