% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetStreetsById}
\alias{databaseGetStreetsById}
\title{Возвращает информацию об улицах по их идентификаторам}
\usage{
databaseGetStreetsById(street_ids = "")
}
\arguments{
\item{street_ids}{Идентификаторы улиц}
}
\description{
Возвращает информацию об улицах по их идентификаторам
}
\examples{
\dontrun{
databaseGetStreetsById(1)
}
}

