% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetStreetsById}
\alias{databaseGetStreetsById}
\title{Returns information about streets by their IDs}
\usage{
databaseGetStreetsById(street_ids = "")
}
\arguments{
\item{street_ids}{Street IDs.}
}
\value{
Returns a list of street objects, each contains id and title fields.
}
\description{
Returns information about streets by their IDs
}
\examples{
databaseGetStreetsById(1)
}

