% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databaseGetUniversities}
\alias{databaseGetUniversities}
\title{Returns a list of higher education institutions}
\usage{
databaseGetUniversities(q = "", country_id = "", city_id = "",
  offset = "", count = "100")
}
\arguments{
\item{q}{Search query.}

\item{country_id}{Country ID.}

\item{city_id}{City ID.}

\item{offset}{Offset needed to return a specific subset of universities.}

\item{count}{Number of universities to return.}
}
\value{
Returns an array of objects with id and title fields.
}
\description{
Returns a list of higher education institutions
}
\examples{
databaseGetUniversities(city_id = '2')
}

