% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{databaseGetUniversities}
\alias{databaseGetUniversities}
\title{Returns a list of higher education institutions}
\usage{
databaseGetUniversities(q = "", country_id = "", city_id = "",
  offset = "", count = "100", v = getAPIVersion())
}
\arguments{
\item{q}{Search query}

\item{country_id}{Country ID}

\item{city_id}{City ID}

\item{offset}{Offset needed to return a specific subset of universities}

\item{count}{Number of universities to return}

\item{v}{Version of API}
}
\description{
Returns a list of higher education institutions
}
\examples{
\dontrun{
databaseGetUniversities(city_id = '2')
}
}

