% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{db_get_connection}
\alias{db_get_connection}
\title{Get a mongo connection object}
\usage{
db_get_connection(collection_name, collection_suffix = "",
  db_name = db_getActive())
}
\arguments{
\item{collection_name}{Collection name}

\item{collection_suffix}{Collection suffix}

\item{db_name}{Database name}
}
\description{
Get a mongo connection object
}

