% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_connection.R
\name{db_save}
\alias{db_save}
\title{Save object to db}
\usage{
db_save(object, collection, suffix = "", db_name = db_getActive())
}
\arguments{
\item{object}{Object to save}

\item{collection}{Collection name}

\item{suffix}{Collection suffix}

\item{db_name}{Database name}
}
\description{
Save object to db
}

