% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2orgtable.R
\name{df2orgtable}
\alias{df2orgtable}
\title{Convert a data frame to an org table and print it}
\usage{
df2orgtable(df, colorizep = FALSE, thold = 0.05, rownames = FALSE)
}
\arguments{
\item{df}{Name of a data element in your working environment}

\item{colorizep}{When set to 'auto', it will find columns ending in 'p' or
'fdr' and color the values smaller than the threshold}

\item{thold}{Threshold for coloring}

\item{rownames}{If set to TRUE, print rownames in output table. If set to
a string, use this as the header of the rownames column and include it in the
output}
}
\description{
Convert a data frame to an org table and print it
}
