% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2orgtable.R
\name{df2orgtable}
\alias{df2orgtable}
\title{Convert a data frame to an org table and print it}
\usage{
df2orgtable(df)
}
\arguments{
\item{df}{Name of a data element in your working environment}
}
\description{
Convert a data frame to an org table and print it
}
