% Generated by roxygen2 (4.0.2): do not edit by hand
\name{execute}
\alias{execute}
\title{Универсальный метод, который позволяет запускать последовательность других методов, сохраняя и фильтруя промежуточные результаты}
\usage{
execute(code)
}
\description{
Универсальный метод, который позволяет запускать последовательность других методов, сохраняя и фильтруя промежуточные результаты
}

