% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{execute}
\alias{execute}
\title{A universal method for calling a sequence of other methods while saving and filtering interim results}
\usage{
execute(code, params = list())
}
\arguments{
\item{code}{Algorithm code in VKScript}

\item{params}{Parameters list}
}
\description{
A universal method for calling a sequence of other methods while saving and filtering interim results
}

