% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{getArbitraryNetwork}
\alias{getArbitraryNetwork}
\title{Building a friend graph for an arbitrary list of users}
\usage{
getArbitraryNetwork(users_ids, format = "edgelist")
}
\arguments{
\item{users_ids}{User IDs}

\item{format}{Either "edgelist" for a list of edges or "adjmatrix" for an adjacency matrix}
}
\description{
Building a friend graph for an arbitrary list of users
}

