% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{getArbitraryNetwork}
\alias{getArbitraryNetwork}
\title{Building a friend graph for an arbitrary list of users}
\usage{
getArbitraryNetwork(users_ids)
}
\arguments{
\item{users_ids}{User IDs}
}
\description{
Building a friend graph for an arbitrary list of users
}

