% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getArbitraryNetwork}
\alias{getArbitraryNetwork}
\title{Создание графа друзей для произвольного списка пользователей}
\usage{
getArbitraryNetwork(users_ids)
}
\arguments{
\item{users_ids}{Произвольный список пользователей, по которым требуется построить граф друзей}
}
\description{
Создание графа друзей для произвольного списка пользователей
}

