% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFriends}
\alias{getFriends}
\title{Возвращает список идентификаторов друзей пользователя или расшиернную информацию о друзьях пользователя}
\usage{
getFriends(user_id = "", order = "", list_id = "", count = "",
  offset = "", fields = "", name_case = "", v = getAPIVersion())
}
\arguments{
\item{user_id}{Идентификатор пользователя, для которого необходимо получить список друзей}

\item{order}{Порядок, в котором нужно вернуть список друзей}

\item{list_id}{Идентификатор списка друзей}

\item{count}{Количество друзей, которое необходимо вернуть}

\item{offset}{Смещение, необходимое для выборки определенного подмножества друзей}

\item{fields}{Список дополнительных полей, которые необходимо вернуть}

\item{name_case}{Падеж для склонения имени и фамилии пользователя}

\item{v}{Версия API}
}
\description{
Возвращает список идентификаторов друзей пользователя или расшиернную информацию о друзьях пользователя
}
\examples{
\dontrun{
friends_list <- getFriends(user_id='1', order='name', fields='bdate')
friends <- friends_list$items
}
}

