% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{getFriendsBy25}
\alias{getFriendsBy25}
\title{Returns a list of friends IDs for the specified users}
\usage{
getFriendsBy25(user_ids, v = getAPIVersion())
}
\arguments{
\item{user_ids}{User IDs (maximum 25)}

\item{v}{Version of API}
}
\description{
Returns a list of friends IDs for the specified users
}
\examples{
\dontrun{
my_friends <- getFriends()
friends_of_friends <- getFriendsBy25(my_friends$items[1:25])
}
}

