% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFriendsBy25}
\alias{getFriendsBy25}
\title{Returns a list of friends IDs for the specified users}
\usage{
getFriendsBy25(user_ids)
}
\arguments{
\item{user_ids}{User IDs}
}
\description{
Returns a list of friends IDs for the specified users
}

