% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFriendsFor}
\alias{getFriendsFor}
\title{Returns a list of friends IDs for the specified users}
\usage{
getFriendsFor(users_ids)
}
\arguments{
\item{users_ids}{User IDs}
}
\description{
Returns a list of friends IDs for the specified users
}

