% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGroups}
\alias{getGroups}
\title{Возвращает список сообществ указанного пользователя}
\usage{
getGroups(user_id = "", extended = "", filter = "", fields = "",
  offset = "", count = "", v = "5.29")
}
\arguments{
\item{user_id}{Идентификатор пользователя, информацию о сообществах которого требуется получить}

\item{extended}{Если указать в качестве этого параметра 1, то будет возвращена полная информация о группах пользователя. По умолчанию 0}

\item{filter}{Список фильтров сообществ, которые необходимо вернуть, перечисленные через запятую. Доступны значения admin, editor, moder, groups, publics, events}

\item{fields}{Список дополнительных полей, которые необходимо вернуть}

\item{offset}{Смещение, необходимое для выборки определённого подмножества сообществ}

\item{count}{Количество сообществ, информацию о которых нужно вернуть}

\item{v}{Версия API}
}
\description{
Возвращает список сообществ указанного пользователя
}
\examples{
groups <- getGroups('1', fields='sex,bdate,city,country,photo_50,education,interests,music,movies,tv,books,games,about,quotes,personal')
}

