% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGroups}
\alias{getGroups}
\title{Возвращает список сообществ указанного пользователя}
\usage{
getGroups(user_id = "", extended = "", filter = "", fields = "",
  offset = "", count = "", v = getAPIVersion())
}
\arguments{
\item{user_id}{Идентификатор пользователя, информацию о сообществах которого требуется получить}

\item{extended}{Если указать в качестве этого параметра 1, то будет возвращена полная информация о группах пользователя. По умолчанию 0}

\item{filter}{Список фильтров сообществ, которые необходимо вернуть, перечисленные через запятую. Доступны значения admin, editor, moder, groups, publics, events}

\item{fields}{Список дополнительных полей, которые необходимо вернуть}

\item{offset}{Смещение, необходимое для выборки определённого подмножества сообществ}

\item{count}{Количество сообществ, информацию о которых нужно вернуть}

\item{v}{Версия API}
}
\description{
Возвращает список сообществ указанного пользователя
}
\examples{
groups <- getGroups('1', fields='city,country,place,description,wiki_page,members_count,counters,start_date,finish_date,can_post,can_see_all_posts,activity,status,contacts,links,fixed_post,verified,site,can_create_topic')
}

