% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGroups}
\alias{getGroups}
\title{Returns a list of the communities to which a user belongs}
\usage{
getGroups(user_id = "", extended = "", filter = "", fields = "",
  offset = "", count = "", v = getAPIVersion())
}
\arguments{
\item{user_id}{User ID}

\item{extended}{1 — to return complete information about a user's communities; 0 — to return a list of community IDs without any additional fields (default)}

\item{filter}{Types of communities to return: admin, editor, moder, groups, publics, events}

\item{fields}{Profile fields to return}

\item{offset}{Offset needed to return a specific subset of communities}

\item{count}{Number of communities to return (maximum value 1000)}

\item{v}{Version of API}
}
\description{
Returns a list of the communities to which a user belongs
}
\examples{
\dontrun{
groups <- getGroups('1', fields='city,country,place,description,wiki_page,members_count,counters,start_date,finish_date,can_post,can_see_all_posts,activity,status,contacts,links,fixed_post,verified,site,can_create_topic')
}
}

