% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGroupsMembers}
\alias{getGroupsMembers}
\title{Returns a list of community members}
\usage{
getGroupsMembers(group_id = "", sort = "", offset = "", count = "",
  fields = "", filter = "")
}
\arguments{
\item{group_id}{ID or screen name of the community.}

\item{sort}{Sort order. Available values: id_asc, id_desc, time_asc, time_desc. time_asc and time_desc are availavle only if the method is called by the group's moderator.}

\item{offset}{Offset needed to return a specific subset of community members.}

\item{count}{Number of community members to return.}

\item{fields}{List of additional fields to be returned.}

\item{filter}{\itemize{
\item \strong{friends} – only friends in this community will be returned;
\item \strong{unsure} – only those who pressed 'I may attend' will be returned (if it's an event).
}}
}
\description{
Returns a list of community members
}
\examples{
groups <- getGroups('1', fields='sex,bdate,city,country,photo_50,education,interests,music,movies,tv,books,games,about,quotes,personal')
}

