% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGroupsMembers}
\alias{getGroupsMembers}
\title{Возвращает список участников сообщества}
\usage{
getGroupsMembers(group_id = "", sort = "", offset = "", count = "",
  fields = "", filter = "", v = getAPIVersion())
}
\arguments{
\item{group_id}{Идентификатор пользователя, информацию о сообществах которого требуется получить}

\item{sort}{Если указать в качестве этого параметра 1, то будет возвращена полная информация о группах пользователя. По умолчанию 0}

\item{offset}{Смещение, необходимое для выборки определённого подмножества сообществ}

\item{count}{Количество сообществ, информацию о которых нужно вернуть}

\item{fields}{Список дополнительных полей, которые необходимо вернуть}

\item{filter}{Список фильтров сообществ, которые необходимо вернуть, перечисленные через запятую. Доступны значения admin, editor, moder, groups, publics, events}

\item{v}{Версия API}
}
\description{
Возвращает список участников сообщества
}
\examples{
groups <- getGroupsMembers('1', fields='sex,bdate,city,country,photo_50,education,interests,music,movies,tv,books,games,about,quotes,personal')
}

