% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{getGroupsMembers}
\alias{getGroupsMembers}
\title{Returns a list of community members}
\usage{
getGroupsMembers(
  group_id = "",
  sort = "",
  offset = "",
  count = "",
  fields = "",
  filter = "",
  v = getAPIVersion()
)
}
\arguments{
\item{group_id}{ID or screen name of the community}

\item{sort}{Sort order}

\item{offset}{Offset needed to return a specific subset of community members}

\item{count}{Number of community members to return (maximum value 1000)}

\item{fields}{List of additional fields to be returned}

\item{filter}{friends - only friends in this community will be returned; unsure - only those who pressed 'I may attend' will be returned (if it's an event)}

\item{v}{Version of API}
}
\description{
Returns a list of community members
}
\examples{
\dontrun{
members <- getGroupsMembers(1, fields='sex,bdate,city')
}
}
