% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGroupsMembersExecute}
\alias{getGroupsMembersExecute}
\title{Возвращает список всех участников сообщества}
\usage{
getGroupsMembersExecute(group_id = "", fields = "", filter = "",
  v = getAPIVersion())
}
\arguments{
\item{group_id}{Идентификатор или короткое имя сообщества}

\item{fields}{Список дополнительных полей, которые необходимо вернуть}

\item{filter}{Доступны значения friends, unsure, managers}

\item{v}{Версия API}
}
\description{
Возвращает список всех участников сообщества
}

