% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMutual}
\alias{getMutual}
\title{Возвращает список идентификаторов общих друзей между парой пользователей}
\usage{
getMutual(source_id = "", target_uid = "", target_uids = "", order = "",
  count = "", offset = "", v = getAPIVersion())
}
\arguments{
\item{source_id}{Идентификатор пользователя, чьи друзья пересекаются с друзьями пользователя с идентификатором target_uid}

\item{target_uid}{Идентификатор пользователя, с которым необходимо искать общих друзей}

\item{target_uids}{Список идентификаторов пользователей, с которыми необходимо искать общих друзей}

\item{order}{Порядок, в котором нужно вернуть список общих друзей}

\item{count}{Количество общих друзей, которое нужно вернуть}

\item{offset}{Смещение, необходимое для выборки определенного подмножества общих друзей}

\item{v}{Версия API}
}
\description{
Возвращает список идентификаторов общих друзей между парой пользователей
}
\examples{
\dontrun{
mutual_friends <- getMutual(target_uid='1')
}
}

