% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getUsers}
\alias{getUsers}
\title{Returns detailed information on users}
\usage{
getUsers(user_ids = "", fields = "", name_case = "",
  v = getAPIVersion())
}
\arguments{
\item{user_ids}{User IDs or screen names (screen_name). By default, current user ID}

\item{fields}{Profile fields to return}

\item{name_case}{Case for declension of user name and surname}

\item{v}{Version of API}
}
\description{
Returns detailed information on users
}
\examples{
\dontrun{
user <- getUsers('1', fields='sex,bdate,city,country,photo_50,education,interests,music,movies,tv,books,games,about,quotes,personal')
}
}

