% Generated by roxygen2 (4.0.2): do not edit by hand
\name{likesGetList}
\alias{likesGetList}
\title{Returns a list of IDs of users who added the specified object to their Likes list}
\usage{
likesGetList(type = "", owner_id = "", item_id = "", page_url = "",
  filter = "", friends_only = "0", extended = "", offset = "",
  count = "100")
}
\arguments{
\item{type}{Object type:
\itemize{
\item \strong{post} — post on user or community wall;
\item \strong{comment} — comment on a wall post;
\item \strong{photo} — photo;
\item \strong{audio} — audio;
\item \strong{video} — video;
\item \strong{note} — note;
\item \strong{photo_comment} — comment on the photo;
\item \strong{video_comment} — comment on the video;
\item \strong{topic_comment} — comment in the discussion;
\item \strong{sitepage} — page of the site where the Like widget is installed.
}}

\item{owner_id}{ID of the user, community, or application that owns the object. If the type parameter is set as sitepage, the application ID is passed as owner_id. If the type parameter is not set, the owner_id is assumed to be either the current user or the same application ID as if the type parameter was set to sitepage.}

\item{item_id}{Object ID. If type is set as sitepage, item_id can include the page_id parameter value used during initialization of the Like widget.}

\item{page_url}{URL of the page where the Like widget is installed. Used instead of the item_id parameter.}

\item{filter}{Filters to apply:
\itemize{
\item \strong{likes} — returns information about all users who liked the object (default);
\item \strong{copies} — returns information only about users who told their friends about the object.
}}

\item{friends_only}{Specifies which users are returned:
\itemize{
\item 1 — to return only the current user's friends;
\item 0 — to return all users (default).
}}

\item{extended}{Specifies whether extended information will be returned.
\itemize{
\item 1 — to return extended information about users and communities from the Likes list;
\item 0 — to return no additional information (default).
}}

\item{offset}{Offset needed to select a specific subset of users.}

\item{count}{Number of user IDs to return (maximum 1000). Default is 100 if friends_only is set to 0; otherwise, the default is 10 if friends_only is set to 1.}
}
\description{
Returns a list of IDs of users who added the specified object to their Likes list
}

