% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likes.R
\name{likesGetList}
\alias{likesGetList}
\title{Returns a list of IDs of users who added the specified object to their Likes list}
\usage{
likesGetList(
  type = "",
  owner_id = "",
  item_id = "",
  page_url = "",
  filter = "",
  friends_only = "0",
  extended = "",
  offset = "",
  count = "100",
  skip_own = 0,
  v = getAPIVersion()
)
}
\arguments{
\item{type}{Object type}

\item{owner_id}{ID of the user, community, or application that owns the object}

\item{item_id}{Object ID}

\item{page_url}{URL of the page where the Like widget is installed. Used instead of the item_id parameter}

\item{filter}{Filters to apply: likes - returns information about all users who liked the object (default); copies - returns information only about users who told their friends about the object}

\item{friends_only}{Specifies which users are returned: 1 - to return only the current user's friends; 0 - to return all users (default)}

\item{extended}{Specifies whether extended information will be returned. 1 - to return extended information about users and communities from the Likes list; 0 - to return no additional information (default)}

\item{offset}{Offset needed to select a specific subset of users}

\item{count}{Number of user IDs to return (maximum 1000)}

\item{skip_own}{Flag, either 1 or 0}

\item{v}{Version of API}
}
\description{
Returns a list of IDs of users who added the specified object to their Likes list
}
