% Generated by roxygen2 (4.0.2): do not edit by hand
\name{messagesGet}
\alias{messagesGet}
\title{Возвращает список входящих либо исходящих личных сообщений текущего пользователя}
\usage{
messagesGet(out = "", offset = "", count = "", time_offset = "",
  filters = "", preview_length = "", last_message_id = "",
  v = getAPIVersion())
}
\arguments{
\item{out}{Если этот параметр равен 1, сервер вернет исходящие сообщения}

\item{offset}{Смещение, необходимое для выборки определенного подмножества сообщений}

\item{count}{Количество сообщений, которое необходимо получить}

\item{time_offset}{Максимальное время, прошедшее с момента отправки сообщения до текущего момента в секундах. 0, если Вы хотите получить сообщения любой давности}

\item{filters}{Фильтр возвращаемых сообщений: 8 — важные сообщения}

\item{preview_length}{Количество символов, по которому нужно обрезать сообщение. Укажите 0, если Вы не хотите обрезать сообщение}

\item{last_message_id}{Идентификатор сообщения, полученного перед тем, которое нужно вернуть последним}
}
\description{
Возвращает список входящих либо исходящих личных сообщений текущего пользователя
}

