% Generated by roxygen2 (4.0.2): do not edit by hand
\name{messagesGet}
\alias{messagesGet}
\title{Returns a list of the current user's incoming or outgoing private messages}
\usage{
messagesGet(out = "", offset = "", count = "", time_offset = "",
  filters = "", preview_length = "", last_message_id = "")
}
\arguments{
\item{out}{1 — to return outgoing messages, 0 — to return incoming messages (default).}

\item{offset}{Offset needed to return a specific subset of messages.}

\item{count}{Number of messages to return (default 20, maximum value 200).}

\item{time_offset}{Maximum time since a message was sent, in seconds. To return messages without a time limitation, set as 0.}

\item{filters}{Filter to apply:
\itemize{
\item 1 — unread only;
\item 2 — not from the chat;
\item 4 — messages from friends.
}
If the 4 flag is set, the 1 and 2 flags are not considered.}

\item{preview_length}{Number of characters after which to truncate a previewed message. To preview the full message, specify 0.}

\item{last_message_id}{ID of the message received before the message that will be returned last (provided that no more than count messages were received before it; otherwise offset parameter shall be used).}
}
\value{
Returns a list of message objects.
}
\description{
Returns a list of the current user's incoming or outgoing private messages
}

