% Generated by roxygen2 (4.0.2): do not edit by hand
\name{messagesGetHistory}
\alias{messagesGetHistory}
\title{Returns message history for the specified user or group chat}
\usage{
messagesGetHistory(offset = "", count = "", user_id = "", chat_id = "",
  start_message_id = "", rev = "")
}
\arguments{
\item{offset}{Offset needed to return a specific subset of messages.}

\item{count}{Number of messages to return (default 20, maximum value 200).}

\item{user_id}{ID of the user whose message history you want to return.}

\item{chat_id}{Chat ID}

\item{start_message_id}{Starting message ID from which to return history.}

\item{rev}{Sort order:
\itemize{
\item 1 — return messages in chronological order;
\item 0 — return messages in reverse chronological order.
}}
}
\value{
Returns a list of message objects.
}
\description{
Returns message history for the specified user or group chat
}

