% Generated by roxygen2 (4.0.2): do not edit by hand
\name{messagesGetHistory}
\alias{messagesGetHistory}
\title{Returns message history for the specified user or group chat}
\usage{
messagesGetHistory(offset = "", count = "", user_id = "", peer_id = "",
  start_message_id = "", rev = "", v = getAPIVersion())
}
\arguments{
\item{offset}{Offset needed to return a specific subset of messages}

\item{count}{Number of messages to return (maximum value 200)}

\item{user_id}{ID of the user whose message history you want to return}

\item{peer_id}{}

\item{start_message_id}{Starting message ID from which to return history}

\item{rev}{Sort order: 1 — return messages in chronological order; 0 — return messages in reverse chronological order}

\item{v}{Version of API}
}
\description{
Returns message history for the specified user or group chat
}

