% Generated by roxygen2 (4.0.2): do not edit by hand
\name{messagesGetHistoryAll}
\alias{messagesGetHistoryAll}
\title{Возвращает историю всех сообщений для указанного пользователя}
\usage{
messagesGetHistoryAll(user_id)
}
\arguments{
\item{user_id}{Идентификатор пользователя, историю переписки с которым необходимо вернуть
messagesGetAll()}
}
\description{
Возвращает историю всех сообщений для указанного пользователя
}
\examples{
\dontrun{
friends <- getFriends(fields = 'sex')$items
friends_ids <- friends$id
messages_for_friend <- data.frame()
for (friend_id in friends_ids) {
   messages_for_friend <- rbind(messages_for_friend, messagesGetHistoryAll(friend_id))
}
}
}

