% Generated by roxygen2 (4.0.2): do not edit by hand
\name{messagesGetHistoryAll}
\alias{messagesGetHistoryAll}
\title{Returns all message history for the specified user or group chat}
\usage{
messagesGetHistoryAll(user_id)
}
\arguments{
\item{user_id}{ID of the user whose message history you want to return
messagesGetAll()}
}
\description{
Returns all message history for the specified user or group chat
}
\examples{
\dontrun{
friends <- getFriends(fields = 'sex')$items
friends_ids <- friends$id
messages_for_friend <- data.frame()
for (friend_id in friends_ids) {
   messages_for_friend <- rbind(messages_for_friend, messagesGetHistoryAll(friend_id))
}
}
}

