% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{messagesSplitByDate}
\alias{messagesSplitByDate}
\title{Split messages by days, weeks, months}
\usage{
messagesSplitByDate(messages, format = "\%y-\%m-\%d")
}
\arguments{
\item{messages}{List of messages from messagesGet()}

\item{format}{Character string giving a date-time format as used by strptime}
}
\description{
Split messages by days, weeks, months
}
