% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queryBuilder}
\alias{queryBuilder}
\title{Returns a query string}
\usage{
queryBuilder(method_name, ...)
}
\arguments{
\item{method_name}{Method name}

\item{...}{Method arguments}
}
\description{
Returns a query string
}
