% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search.getHints}
\alias{search.getHints}
\title{Allows the programmer to do a quick search for any substring}
\usage{
search.getHints(q = "", limit = "", filters = "", search_global = "",
  v = getAPIVersion())
}
\arguments{
\item{q}{Search query string}

\item{limit}{Maximum number of results to return}

\item{filters}{}

\item{search_global}{}

\item{v}{Version of API}
}
\description{
Allows the programmer to do a quick search for any substring
}

