% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search.getHints}
\alias{search.getHints}
\title{Позволяет получить результаты быстрого поиска по произвольной подстроке}
\usage{
search.getHints(q = "", limit = "", filters = "", search_global = "",
  v = getAPIVersion())
}
\arguments{
\item{q}{Текст запроса, результаты которого нужно получить}

\item{limit}{Ограничение на количество возвращаемых результатов}

\item{filters}{Перечисленные через запятую типы данных, которые необходимо вернуть}

\item{search_global}{По умолчанию к результатам поиска добавляются результаты глобального поиска по всем пользователям и группам, это можно отключить передав 0}
}
\description{
Позволяет получить результаты быстрого поиска по произвольной подстроке
}

