% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search.getHints}
\alias{search.getHints}
\title{Allows the programmer to do a quick search for any substring}
\usage{
search.getHints(
  q = "",
  limit = "",
  filters = "",
  search_global = "",
  v = getAPIVersion()
)
}
\arguments{
\item{q}{Search query string}

\item{limit}{Maximum number of results to return}

\item{filters}{List of comma-separated words}

\item{search_global}{Flag, either 1 or 0, default 1}

\item{v}{Version of API}
}
\description{
Allows the programmer to do a quick search for any substring
}
