% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wallGetById}
\alias{wallGetById}
\title{Возвращает список записей со стен пользователей или сообществ по их идентификаторам}
\usage{
wallGetById(posts = "", extended = "", copy_history_depth = "")
}
\arguments{
\item{posts}{Перечисленные через запятую идентификаторы, которые представляют собой идущие через знак подчеркивания id владельцев стен и id самих записей на стене}

\item{extended}{1 - возвращает объекты пользователей и групп, необходимые для отображения записей}

\item{copy_history_depth}{Определяет размер массива copy_history, возвращаемого в ответе, если запись является репостом записи с другой стены}
}
\description{
Возвращает список записей со стен пользователей или сообществ по их идентификаторам
}

