% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wallGetComments}
\alias{wallGetComments}
\title{Returns a list of comments on a post on a user wall or community wall}
\usage{
wallGetComments(owner_id = "", post_id = "", need_likes = "",
  offset = "", count = "10", sort = "", preview_length = "0",
  extended = "")
}
\arguments{
\item{owner_id}{User ID or community ID. Use a negative value to designate a community ID.}

\item{post_id}{Post ID.}

\item{need_likes}{1 — to return the likes field, 0 — not to return the likes field (default).}

\item{offset}{Offset needed to return a specific subset of comments.}

\item{count}{Number of comments to return (maximum 100).}

\item{sort}{Sort order: asc — chronological, desc — reverse chronological.}

\item{preview_length}{Number of characters at which to truncate comments when previewed. By default, 90. Specify 0 if you do not want to truncate comments.}

\item{extended}{Flag, either 1 or 0.}
}
\description{
Returns a list of comments on a post on a user wall or community wall
}

