% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wallSearch}
\alias{wallSearch}
\title{Метод, позволяющий осуществлять поиск по стенам пользователей}
\usage{
wallSearch(owner_id = "", domain = "", query = "", owners_only = "",
  count = "20", offset = "0", extended = "")
}
\arguments{
\item{owner_id}{Идентификатор пользователя или сообщества}

\item{domain}{Короткий адрес пользователя или сообщества}

\item{query}{Поисковой запрос}

\item{owners_only}{1 — возвращать только записи от имени владельца стены}

\item{count}{Количество записей, которые необходимо вернуть}

\item{offset}{Смещение}

\item{extended}{Возвращать ли расширенную информацию о записях}
}
\description{
Метод, позволяющий осуществлять поиск по стенам пользователей
}

