% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wall.R
\name{wallSearch}
\alias{wallSearch}
\title{Allows to search posts on user or community walls}
\usage{
wallSearch(
  owner_id = "",
  domain = "",
  query = "",
  owners_only = "",
  count = "20",
  offset = "0",
  extended = "",
  fields = "",
  v = getAPIVersion()
)
}
\arguments{
\item{owner_id}{User or community id. Remember that for a community owner_id must be negative.}

\item{domain}{User or community screen name.}

\item{query}{Search query string.}

\item{owners_only}{1 - returns only page owner's posts.}

\item{count}{Count of posts to return.}

\item{offset}{Results offset.}

\item{extended}{Show extended post info.}

\item{fields}{List of comma-separated words}

\item{v}{Version of API}
}
\value{
If executed successfully, returns a list of post objects.
}
\description{
Allows to search posts on user or community walls
}
